refactor: split MVP components
